添加声音 fix：开始战斗语音聊天失效 fix：结束面板数据显示错误 fix：结束时 玩家豆子未在ui中惊醒更新

+++ Protocol/Dto/Constant/CardConstant.cs
@@ -183,7 +183,7 @@ namespace Protocol.Dto.Constant
                     {
                         if (IsDouble(cards))
                         {
-                            type = CardsType.Single;
+                            type = CardsType.Double;
                         }
                         else if (IsJokerBoom(cards))
                         {
